New Automotive definitions
This replaces the old Automotive definitions.  The old A-Team/Toolchain
workloads were unused and obsolete.  The previous basic workloads have
now been merged into the standard in-vehicle environment.  The
development workload was renamed to off-vehicle set.  The Neptune and
infotainment workloads were adjusted to this new Automotive schema.

Confirmed with all parties.

Signed-off-by: Petr Šabata <contyk@redhat.com>